+++ json-path/src/main/java/com/jayway/jsonpath/Filter.java
@@ -14,31 +14,17 @@
  */
 package com.jayway.jsonpath;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
+import java.util.Stack;
+import java.util.regex.Pattern;
 
 /**
  *
  */
-public class Filter implements Predicate {
+public abstract class Filter implements Predicate {
 
-    protected final List<Predicate> criteriaList;
-
-    private Filter() {
-        criteriaList = Collections.emptyList();
-    }
-
-    private Filter(Predicate criteria) {
-        criteriaList = Collections.singletonList(criteria);
-    }
-
-    private Filter(List<Predicate> criteriaList) {
-        this.criteriaList = new ArrayList<Predicate>(criteriaList);
-    }
-
-
+    private static final Pattern OPERATOR_SPLIT = Pattern.compile("((?<=&&|\\|\\|)|(?=&&|\\|\\|))");
+    private static final String AND = "&&";
+    private static final String OR = "||";
 
     /**
      * Creates a new Filter based on given criteria
@@ -46,45 +32,63 @@ public class Filter implements Predicate {
      * @return a new Filter
      */
     public static Filter filter(Predicate criteria) {
-        return new Filter(criteria);
+        return new SingleFilter(criteria);
     }
 
-    /**
-     * Create a new Filter based on given list of criteria.
-     * @param criteriaList list of criteria
-     * @return
-     */
-    public static Filter filter(List<Predicate> criteriaList) {
-        return new Filter(criteriaList);
-    }
 
     @Override
-    public boolean apply(PredicateContext ctx) {
-        for (Predicate criteria : criteriaList) {
-            if (!criteria.apply(ctx)) {
-                return false;
-            }
-        }
-        return true;
-    }
+    public abstract boolean apply(PredicateContext ctx);
 
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        for (Predicate crit : criteriaList) {
-            sb.append(crit.toString());
-        }
-        return sb.toString();
-    }
 
     public Filter or(final Predicate other){
         return new OrFilter(this, other);
     }
 
     public Filter and(final Predicate other){
-        return filter(Arrays.asList(this, other));
+        return new AndFilter(this, other);
     }
-    
+
+    private static final class SingleFilter extends Filter {
+
+        private final Predicate predicate;
+
+        private SingleFilter(Predicate predicate) {
+            this.predicate = predicate;
+        }
+
+        @Override
+        public boolean apply(PredicateContext ctx) {
+            return predicate.apply(ctx);
+        }
+
+        @Override
+        public String toString() {
+            return predicate.toString();
+        }
+    }
+
+    private static final class AndFilter extends Filter {
+
+        private final Predicate left;
+        private final Predicate right;
+
+        private AndFilter(Predicate left, Predicate right) {
+            this.left = left;
+            this.right = right;
+        }
+
+        @Override
+        public boolean apply(PredicateContext ctx) {
+            boolean a = left.apply(ctx);
+            return a && right.apply(ctx);
+        }
+
+        @Override
+        public String toString() {
+            return left.toString() + " && " + right.toString();
+        }
+    }
+
     private static final class OrFilter extends Filter {
 
         private final Predicate left;
@@ -94,11 +98,84 @@ public class Filter implements Predicate {
             this.left = left;
             this.right = right;
         }
-  
+
+        public Filter and(final Predicate other){
+            return new OrFilter(left, new AndFilter(right, other));
+        }
+
         @Override
         public boolean apply(PredicateContext ctx) {
             boolean a = left.apply(ctx);
             return a || right.apply(ctx);
         }
+
+        @Override
+        public String toString() {
+            return left.toString() + " || " + right.toString();
+        }
+    }
+
+
+    public static Filter parse(String filter){
+        filter = filter.trim();
+        if(!filter.startsWith("[") || !filter.endsWith("]")){
+            throw new InvalidPathException("Filter must start with '[' and end with ']'. " + filter);
+        }
+        filter = filter.substring(1, filter.length()-1).trim();
+        if(!filter.startsWith("?")){
+            throw new InvalidPathException("Filter must start with '[?' and end with ']'. " + filter);
+        }
+        filter = filter.substring(1).trim();
+        if(!filter.startsWith("(") || !filter.endsWith(")")){
+            throw new InvalidPathException("Filter must start with '[?(' and end with ')]'. " + filter);
+        }
+        filter = filter.substring(1, filter.length()-1).trim();
+
+        String[] split = OPERATOR_SPLIT.split(filter);
+        Stack<String> operators = new Stack<String>();
+        Stack<Criteria> criteria = new Stack<Criteria>();
+
+        for (String exp : split) {
+            exp = exp.trim();
+            if(AND.equals(exp) || OR.equals(exp)){
+                operators.push(exp);
+            }
+            else {
+                criteria.push(Criteria.parse(cleanCriteria(exp)));
+            }
+        }
+        Filter root = new SingleFilter(criteria.pop());
+        while(!operators.isEmpty()) {
+            String operator = operators.pop();
+            if (AND.equals(operator)) {
+                root = root.and(criteria.pop());
+
+            } else {
+                if(criteria.isEmpty()){
+                    throw new InvalidPathException("Invalid operators " + filter);
+                }
+                root = root.or(criteria.pop());
+            }
+        }
+        if(!operators.isEmpty() || !criteria.isEmpty()){
+            throw new InvalidPathException("Invalid operators " + filter);
+        }
+        return root;
+    }
+
+    private static String cleanCriteria(String filter){
+        int begin = 0;
+        int end = filter.length() -1;
+
+        char c = filter.charAt(begin);
+        while(c == '[' || c == '?' || c == '(' || c == ' '){
+            c = filter.charAt(++begin);
+        }
+
+        c = filter.charAt(end);
+        while( c == ')' || c == ' '){
+            c = filter.charAt(--end);
+        }
+        return filter.substring(begin, end+1);
     }
 }