+++ changelog.md
@@ -3,6 +3,7 @@ In The Pipe
 * Added EvaluationListener interface that allows abortion of evaluation if criteria is fulfilled.
   this makes it possible to limit the number of results to fetch when a document is scanned. Also 
   added utility method to limit results `JsonPath.parse(json).limit(1).read("$..title", List.class);`
+* Added support for OR in inline filters [?(@.foo == 'bar' || @.foo == 'baz')]  
 
 
 1.1.0 (2014-10-01)